# openapi.json
{
   "openapi":"3.1.0",
   "info":{
      "title":"dianovi Medical Recommendation API",
      "description":"API for patient data and treatment recommendations.",
      "version":"0.1.0"
   },
   "paths":{
      "/api/v1/patients":{
         "post":{
            "tags":[
               "Patients"
            ],
            "summary":"Create Patient",
            "description":"Creates a new patient record in the database.",
            "operationId":"create_patient_api_v1_patients_post",
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/PatientCreate"
                     }
                  }
               }
            },
            "responses":{
               "201":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/Patient"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "get":{
            "tags":[
               "Patients"
            ],
            "summary":"Get All Patients",
            "description":"Retrieves a list of all patients from the database.",
            "operationId":"get_all_patients_api_v1_patients_get",
            "parameters":[
               {
                  "name":"skip",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "type":"integer",
                     "default":0,
                     "title":"Skip"
                  }
               },
               {
                  "name":"limit",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "type":"integer",
                     "default":100,
                     "title":"Limit"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "type":"array",
                           "items":{
                              "$ref":"#/components/schemas/Patient"
                           },
                           "title":"Response Get All Patients Api V1 Patients Get"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/v1/patients/{patient_id}":{
         "get":{
            "tags":[
               "Patients"
            ],
            "summary":"Get Patient",
            "description":"Retrieves the details for a single patient from the database.",
            "operationId":"get_patient_api_v1_patients__patient_id__get",
            "parameters":[
               {
                  "name":"patient_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Patient Id"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/Patient"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "put":{
            "tags":[
               "Patients"
            ],
            "summary":"Update Patient",
            "description":"Updates an existing patient's record.",
            "operationId":"update_patient_api_v1_patients__patient_id__put",
            "parameters":[
               {
                  "name":"patient_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Patient Id"
                  }
               }
            ],
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/PatientCreate"
                     }
                  }
               }
            },
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/Patient"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "delete":{
            "tags":[
               "Patients"
            ],
            "summary":"Delete Patient",
            "description":"Deletes a patient record.",
            "operationId":"delete_patient_api_v1_patients__patient_id__delete",
            "parameters":[
               {
                  "name":"patient_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Patient Id"
                  }
               }
            ],
            "responses":{
               "204":{
                  "description":"Successful Response"
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/v1/patients/{patient_id}/recommendations":{
         "get":{
            "tags":[
               "Patients"
            ],
            "summary":"Get Patient Recommendations",
            "description":"Returns a list of dummy recommendations for a patient.",
            "operationId":"get_patient_recommendations_api_v1_patients__patient_id__recommendations_get",
            "parameters":[
               {
                  "name":"patient_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Patient Id"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{

                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/":{
         "get":{
            "summary":"Read Root",
            "description":"A simple health check endpoint.",
            "operationId":"read_root__get",
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{

                        }
                     }
                  }
               }
            }
         }
      }
   },
   "components":{
      "schemas":{
         "HTTPValidationError":{
            "properties":{
               "detail":{
                  "items":{
                     "$ref":"#/components/schemas/ValidationError"
                  },
                  "type":"array",
                  "title":"Detail"
               }
            },
            "type":"object",
            "title":"HTTPValidationError"
         },
         "Patient":{
            "properties":{
               "patient_id":{
                  "type":"string",
                  "title":"Patient Id"
               },
               "name":{
                  "type":"string",
                  "title":"Name"
               },
               "date_of_birth":{
                  "type":"string",
                  "format":"date",
                  "title":"Date Of Birth"
               }
            },
            "type":"object",
            "required":[
               "patient_id",
               "name",
               "date_of_birth"
            ],
            "title":"Patient"
         },
         "PatientCreate":{
            "properties":{
               "patient_id":{
                  "type":"string",
                  "title":"Patient Id"
               },
               "name":{
                  "type":"string",
                  "title":"Name"
               },
               "date_of_birth":{
                  "type":"string",
                  "format":"date",
                  "title":"Date Of Birth"
               }
            },
            "type":"object",
            "required":[
               "patient_id",
               "name",
               "date_of_birth"
            ],
            "title":"PatientCreate"
         },
         "ValidationError":{
            "properties":{
               "loc":{
                  "items":{
                     "anyOf":[
                        {
                           "type":"string"
                        },
                        {
                           "type":"integer"
                        }
                     ]
                  },
                  "type":"array",
                  "title":"Location"
               },
               "msg":{
                  "type":"string",
                  "title":"Message"
               },
               "type":{
                  "type":"string",
                  "title":"Error Type"
               }
            },
            "type":"object",
            "required":[
               "loc",
               "msg",
               "type"
            ],
            "title":"ValidationError"
         }
      }
   }
}